requirements.txt
openai==1.70.0
tiktoken==0.7.0
streamlit==1.33.0
langchain==0.3.22
langchain-community==0.3.20
langchain-openai==0.3.11
langchain-core==0.3.49
PyMuPDF==1.25.5
faiss-cpu==1.10.0
duckduckgo_search==2025.4.4
html2text==2024.2.26
readability-lxml==0.8.1
lxml_html_clean==0.4.1
pydantic==2.11.1